Improving test for took time
